Add real multi-attachment color render targets to the Vulkan backend
VulkanRenderDevice previously only implemented depth-only framebuffer
attachments - any color attachment request was silently rejected, so
DeferredRenderer's G-buffer pass never worked at all and every
PostEffectsManager-based example rendered into a broken target.

A caller-supplied multi-attachment FrameBuffer arrives as several
independent, self-contained AddAttach() calls, each wrapped in its own
temporary bind/attach/unbind - Vulkan needs every attachment's format
known before building a render pass, so building one on the first
attach call (the old behavior) can't work once there's more than one
attachment. Fixed by accumulating attachments and deferring the real
VkRenderPass/VkFramebuffer build to the FBO's first genuine per-frame
bind, distinguishing that from a self-contained setup-time micro-bind
via an explicit parameter threaded from FrameBuffer.cpp (GL ignores
it - no render-pass concept to defer). Also added the missing
COLOR_ATTACHMENT_BIT usage flag on color textures, and made pipeline
creation target the currently-bound FBO's own render pass instead of
always the swapchain's.

Three more bugs surfaced once color attachments actually started
working, each confirmed via a live regression:
- A single-depth-attachment FBO built through the new general path had
  different subpass-dependency masks than the shared shadow-pipeline
  render pass template, breaking render-pass compatibility for shadow
  casting. Fixed by keeping that one case on the original builder.
- RenderingMesh's pipeline cache was keyed only by shader, not render
  target, so the same mesh drawn into two differently-shaped targets
  (a water reflection FBO and the main swapchain pass) wrongly shared
  one pipeline. Fixed by keying on (shader, current render target) and
  un-coupling pipeline creation from the unrelated VAO cache gate it
  was piggybacking on.
- ForwardRenderer::RenderScene() unconditionally began/ended a
  swapchain frame every time it ran, which broke any example calling
  it more than once per frame for an offscreen pass (water reflection/
  refraction) - the swapchain frame hijacked the command buffer away
  from the offscreen target mid-pass. Fixed by only managing the
  swapchain lifecycle when not already drawing into a caller-bound FBO.

DeferredRendering and IslandDemo both go from broken/crashing to zero
validation errors; full example-suite sweep shows no regressions
elsewhere, GL unaffected throughout.

+++ include/Pyros3D/Rendering/Components/Rendering/RenderingComponent.h
@@ -74,19 +74,29 @@ namespace p3d {
 		// locations can differ across shader variants using this mesh.
 		std::map<uint32, uint32> VAOCache;
 
-		// Vulkan pipeline cache, keyed by shader program - mirrors VAOCache
-		// exactly (see RenderObject()/BindMesh() in IRenderer.cpp). Only
-		// ever populated on the Vulkan backend (GLRenderDevice::CreatePipeline()
-		// exists too, but nothing calls it outside this cache - GL still
-		// uses the individual SetCullFaceMode/SetBlendingEnabled/etc calls
-		// directly). Built from Material's blend/depth/cull/wireframe state
-		// at the moment this (mesh, shader) pair is first bound, not
-		// re-evaluated per object the way GL's own dirty-tracked state is -
-		// correct for any Material whose blend/depth/cull state doesn't
-		// change after the fact for a given mesh/shader pairing (true for
-		// RotatingCube, this backend's only validated target), not a
-		// general solution.
-		std::map<uint32, uint32> PipelineCache;
+		// Vulkan pipeline cache, keyed by (shader program, current render
+		// target) packed into one uint64 - ((uint64)shader << 32) | targetFBO,
+		// see IRenderDevice::GetCurrentRenderTarget(). Mirrors VAOCache's
+		// shader-only keying (see RenderObject()/BindMesh() in IRenderer.cpp)
+		// plus the render-target dimension a Vulkan pipeline needs, since
+		// it bakes in a specific render pass's attachment shape at
+		// creation time - the same mesh+shader drawn into two
+		// differently-shaped targets (e.g. a color-only reflection FBO
+		// versus the main color+depth swapchain pass) needs two separate
+		// pipelines. Only ever populated on the Vulkan backend
+		// (GLRenderDevice::CreatePipeline() exists too, but nothing calls
+		// it outside this cache - GL still uses the individual
+		// SetCullFaceMode/SetBlendingEnabled/etc calls directly, and
+		// GetCurrentRenderTarget() always returns 0 there, collapsing
+		// this back to shader-only keying, matching prior behavior
+		// exactly since GL has no render-pass-shape concept to
+		// disambiguate). Built from Material's blend/depth/cull/wireframe
+		// state at the moment this (mesh, shader, target) triple is first
+		// bound, not re-evaluated per object the way GL's own
+		// dirty-tracked state is - correct for any Material whose
+		// blend/depth/cull state doesn't change after the fact for a
+		// given mesh/shader/target combination.
+		std::map<uint64, uint32> PipelineCache;
 
 		// Materials
 		IMaterial* Material;
